test: Updates project structure
