+++ ios/Tests/Mocks/OktaSdkBridgeMock.swift
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2021-Present, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
+
+import XCTest
+import OktaOidc
+@testable import ReactNativeOktaSdkBridge
+
+final class OktaSdkBridgeMock: OktaSdkBridge {
+    private(set) var eventsRegister: [String: Any] = [:]
+    private var customStateManager: StateManagerProtocol?
+    
+    func setCustomStateManager(_ stateManager: StateManagerProtocol) {
+        self.customStateManager = stateManager
+    }
+    
+    // Must be overriden
+    override func supportedEvents() -> [String]! {
+        super.supportedEvents()
+    }
+    
+    override func sendEvent(withName name: String!, body: Any!) {
+        eventsRegister[name] = body
+    }
+    
+    override func presentedViewController() -> UIViewController? {
+        UIViewController()
+    }
+    
+    override func readStateManager() -> StateManagerProtocol? {
+        if customStateManager != nil {
+            return customStateManager
+        }
+        
+        return config.flatMap {
+            OktaOidcMock.oidcStateManager(with: $0)
+        }
+    }
+}
